[build system] fix build system cache issue
if RTL is changed(from vector compile unit), it will result the same
classpath in elaborate, since the input of elaborate is runClasspath(the
path of classpath directory), it doesn't consume the zinc analysis file.
In the commit, I add upstreamCompileOuput to elaborate dependency, which
pulls all zinc analysis result to this JVM run.