[ci] remove gen-test-elf workflow and rewrite the elf generate script
Artifacts in GitHub action are only retained for 5 days,
which is not ideal to use as a persistant storage.
So this commit remove all the previous implementation
and just let the script compile all the test in local machine.

With nix store, these tests will not compile everytime you enter the shell.

Signed-off-by: Avimitin <dev@avimit.in>