[tests] add memcpy-asm rvv test case
Added relevant test cases for memcpy. Corresponding config files
and asm files have been added, reusing main.S. Since the current
CI only tests for RTL bugs and does not check the correctness of
the case itself, this case will execute to the end regardless of
the result. However, for future correctness testing, I have added
a result at the end of the test that will be convenient for future
maintainers to handle:

- if the test passes, the `a0` register will be set to `0xff`
- if the test fails, the `a0` register will be set to `0xfe`

future maintainers can continue to add assembly code to decide
specifically what to do in case of test pass or fail.

Signed-off-by: Huang Rui <vowstar@gmail.com>